fixed some warnings
