+++ market_economy_simulation/src/ground_plane/ground_plane_mesh.rs
@@ -26,8 +26,7 @@ impl GroundPlaneMesh {
         let quad_size = 1.0 * scale;
         let quad = geometry::Quad::new(quad_size);
 
-        let mut instances: Vec<deferred_color_shader::Instance> = Vec::new();
-        instances.reserve(ground_plane.size());
+        let mut instances: Vec<deferred_color_shader::Instance> = Vec::with_capacity(ground_plane.size());
 
         for y in 0..ground_plane.height() {
             for x in 0..ground_plane.width() {
